server/requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
awsebcli==3.10.0
blessed==1.14.1
botocore==1.5.22
cement==2.8.2
click==6.7
colorama==0.3.7
configparser==3.5.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
Flask==0.12
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.1
MarkupSafe==1.0
mysqlclient==1.3.10
packaging==16.8
passlib==1.7.1
pathspec==0.5.0
pyaml==16.12.2
pyparsing==2.2.0
pyswagger==0.8.27
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
SQLAlchemy==1.1.6
tabulate==0.7.5
termcolor==1.1.0
validate-email==1.3
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.11.15
flasgger-0.5.14